+++ src/pages/spawn/spawnVolume/spawnVolumeTable/spawnVolumeTableActions/mountButton/MountVolumeModal.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from "react";
-import { useMutation, useQuery } from "@apollo/client";
+import React from "react";
+import { useMutation } from "@apollo/client";
 import styled from "@emotion/styled";
 import { Variant } from "@leafygreen-ui/button";
 import { Disclaimer } from "@leafygreen-ui/typography";
@@ -7,17 +7,13 @@ import { Select } from "antd";
 import { Modal } from "components/Modal";
 import { ModalContent, Section, WideButton } from "components/Spawn";
 import { InputLabel } from "components/styles";
-import { pollInterval } from "constants/index";
 import { useBannerDispatchContext } from "context/banners";
 import {
   AttachVolumeToHostMutation,
   AttachVolumeToHostMutationVariables,
-  MyHostsQuery,
-  MyHostsQueryVariables,
 } from "gql/generated/types";
 import { ATTACH_VOLUME } from "gql/mutations/attach-volume";
-import { GET_MY_HOSTS } from "gql/queries";
-import { useNetworkStatus } from "hooks";
+import { useMountVolumeSelect } from "hooks/useMountVolumeSelect";
 import { MyVolume } from "types/spawn";
 
 const { Option } = Select;
@@ -27,10 +23,6 @@ interface Props {
   onCancel: () => void;
   volume: MyVolume;
 }
-interface HostOption {
-  key: string;
-  displayName: string;
-}
 
 export const MountVolumeModal: React.FC<Props> = ({
   visible,
@@ -38,48 +30,13 @@ export const MountVolumeModal: React.FC<Props> = ({
   volume,
 }) => {
   const dispatchBanner = useBannerDispatchContext();
-  const [hostOptions, setHostOptions] = useState<HostOption[]>([]); // dropdown option
-  const [selectedHostId, setSelectedHostId] = useState(""); // selected dropdown item
-  const { data, loading, startPolling, stopPolling } = useQuery<
-    MyHostsQuery,
-    MyHostsQueryVariables
-  >(GET_MY_HOSTS, {
-    pollInterval,
-    onError: (e) => {
-      dispatchBanner.errorBanner(
-        `There was an error loading hosts: ${e.message}`
-      );
-    },
-  });
-  useNetworkStatus(startPolling, stopPolling);
   const targetAvailabilityZone = volume.availabilityZone;
-
-  // set host dropdown options
-  useEffect(() => {
-    if (data?.myHosts) {
-      const opts = data.myHosts
-        // Filter hosts that do not have the same availability zone as the volume.
-        .filter(
-          ({ availabilityZone }) => availabilityZone === targetAvailabilityZone
-        )
-        // Map host to a displayName and ID for the dropdown <Option />
-        .map((host) => ({
-          key: host.id,
-          displayName: host.displayName ? host.displayName : host.id,
-        }))
-        // Sort the dropdown items by display name.
-        .sort((a, b) => a.displayName.localeCompare(b.displayName));
-      setHostOptions(opts);
-    }
-  }, [data, targetAvailabilityZone]);
-
-  // set initially selected host in dropdown
-  useEffect(() => {
-    if (!selectedHostId && hostOptions.length) {
-      setSelectedHostId(hostOptions[0].key);
-    }
-  }, [hostOptions, selectedHostId]);
-
+  const {
+    loading,
+    hostOptions,
+    selectedHostId,
+    setSelectedHostId,
+  } = useMountVolumeSelect({ targetAvailabilityZone });
   const [attachVolume, { loading: loadingAttachVolume }] = useMutation<
     AttachVolumeToHostMutation,
     AttachVolumeToHostMutationVariables
